[codex] Unify streaming planner and CI coverage (#2)
* Unify streaming planner and CI coverage

* Fix GCC numeric literal ambiguity

* Fix Linux GCC build ambiguities

* Fix Linux Python extension linkopts

+++ src/dataflow/core/csv.cc
@@ -30,7 +30,7 @@ bool isDouble(const std::string& s) {
 Value parseCell(const std::string& cell) {
   if (cell.empty()) return Value();
   if (isInt(cell)) {
-    return Value(std::stoll(cell));
+    return Value(static_cast<int64_t>(std::stoll(cell)));
   }
   if (isDouble(cell)) {
     return Value(std::stod(cell));
